evised workflow file with some fixes and clarifications:
Pass variables to Terraform properly: Ensure that the Docker username, password, and GitHub run number are passed to Terraform as arguments in Terraform Apply.
Working Directory for Terraform: Set the working directory for all Terraform commands.
Use of with and env in the Terraform steps: Set env for AWS credentials in terraform apply, and ensure all variables are provided where necessary.